LPD-96185 Extract getTaskItemsActions into a shared util
