+++ modules/apps/site/site-cmp-site-initializer/src/main/resources/META-INF/resources/js/components/props_transformer/views/kanban_view/components/Task.tsx
@@ -7,251 +7,19 @@ import ClayButton from '@clayui/button';
 import Card from '@clayui/card/src/Card';
 import {ClayDropDownWithItems} from '@clayui/drop-down';
 import ClayIcon from '@clayui/icon';
-import {DateRenderer, IItemsActions} from '@liferay/frontend-data-set-web';
+import {DateRenderer} from '@liferay/frontend-data-set-web';
 import {AssigneeAvatar} from '@liferay/object-dynamic-data-mapping-form-field-type';
-import {
-	displayErrorToast,
-	displayRequestSuccessToast,
-} from '@liferay/site-cms-site-initializer';
 import classNames from 'classnames';
-import {navigate} from 'frontend-js-web';
 import React, {forwardRef, useContext} from 'react';
 import {useDrag} from 'react-dnd';
 
-import {
-	deleteTaskById,
-	getUserAccount,
-	patchTaskById,
-	postSubscribeTaskByExternalReferenceCode,
-	postUnsubscribeTaskByExternalReferenceCode,
-} from '../../../../../utils/api';
-import getActionURL from '../../../../../utils/getActionURL';
-import {openCMPModal} from '../../../../../utils/openCMPModal';
-import {
-	displayAssignSuccessToast,
-	displayDeleteSuccessToast,
-} from '../../../../../utils/toastUtil';
+import getTaskItemsActions from '../../../../../utils/getTaskItemsActions';
 import {ITask} from '../../../../../utils/types';
 import StateLabel from '../../../../StateLabel';
-import DeleteTaskModal from '../../../../modal/DeleteTaskModal';
-import EditAssigneeModalContent from '../../../../modal/EditAssigneeModalContent';
 import {KanbanViewContext} from '../context';
 import {ItemTypes} from './Column';
 
 import './Task.scss';
-
-function getTaskItemsActions(
-	itemsActions: IItemsActions[],
-	loadData: Function,
-	task: ITask
-) {
-	const items = [];
-
-	const hasBottomItems = Boolean(task.actions.delete);
-
-	const hasMiddleItems = Boolean(
-		task.actions.assignToMe ||
-			task.actions.subscribe ||
-			task.actions.unsubscribe ||
-			task.actions.update
-	);
-
-	const hasTopItems = Boolean(task.actions.update || task.actions.get);
-
-	if (task.actions.update) {
-		items.push({
-			label: Liferay.Language.get('edit'),
-			onClick: () => {
-				const editURL = getActionURL({
-					actionId: 'edit',
-					itemsActions,
-					task,
-				});
-				if (editURL) {
-					navigate(editURL);
-				}
-			},
-			symbolLeft: 'pencil',
-		});
-	}
-
-	if (task.actions.get) {
-		items.push({
-			label: Liferay.Language.get('view'),
-			onClick: () => {
-				const viewURL = getActionURL({
-					actionId: 'actionLink',
-					itemsActions,
-					task,
-				});
-				if (viewURL) {
-					navigate(viewURL);
-				}
-			},
-			symbolLeft: 'view',
-		});
-	}
-
-	if (hasTopItems && hasMiddleItems) {
-		items.push({
-			type: 'divider' as const,
-		});
-	}
-
-	if (task.actions.subscribe) {
-		items.push({
-			label: Liferay.Language.get('watch-task'),
-			onClick: async () => {
-				const {error} = await postSubscribeTaskByExternalReferenceCode({
-					externalReferenceCode: task.embedded.externalReferenceCode,
-					scopeKey: task.embedded.scopeKey,
-				});
-
-				if (!error) {
-					loadData();
-					displayRequestSuccessToast();
-				}
-				else {
-					displayErrorToast(error);
-				}
-			},
-			symbolLeft: 'bell-on',
-		});
-	}
-
-	if (task.actions.unsubscribe) {
-		items.push({
-			label: Liferay.Language.get('stop-watching-task'),
-			onClick: async () => {
-				const {error} =
-					await postUnsubscribeTaskByExternalReferenceCode({
-						externalReferenceCode:
-							task.embedded.externalReferenceCode,
-						scopeKey: task.embedded.scopeKey,
-					});
-
-				if (!error) {
-					loadData();
-					displayRequestSuccessToast();
-				}
-				else {
-					displayErrorToast(error);
-				}
-			},
-			symbolLeft: 'bell-off',
-		});
-	}
-
-	if (task.actions.assignToMe) {
-		items.push({
-			label: Liferay.Language.get('assign-to-me'),
-			onClick: async () => {
-				const user = (await getUserAccount(
-					Liferay.ThemeDisplay.getUserId().toString()
-				)) as {
-					externalReferenceCode: string;
-					name: string;
-				};
-
-				const {error} = await patchTaskById({
-					body: {
-						assignTo: {
-							externalReferenceCode: user.externalReferenceCode,
-							name: user.name,
-							type: 'User',
-						},
-					},
-					taskId: String(task.embedded.id),
-				});
-
-				if (!error) {
-					loadData();
-					displayAssignSuccessToast(task.embedded.title, user.name);
-				}
-				else {
-					displayErrorToast(error);
-				}
-			},
-		});
-	}
-
-	if (task.actions.update) {
-		items.push({
-			label: Liferay.Language.get('assign-to-...'),
-			onClick: async () => {
-				await openCMPModal({
-					center: true,
-					contentComponent: ({
-						closeModal,
-					}: {
-						closeModal: () => void;
-					}) => (
-						<EditAssigneeModalContent
-							closeModal={closeModal}
-							loadData={loadData}
-							taskId={String(task.embedded.id)}
-							taskTitle={task.embedded.title}
-							value={task.embedded.assignTo}
-						/>
-					),
-					size: 'md',
-				});
-			},
-		});
-	}
-
-	if (hasBottomItems) {
-		if (hasTopItems || hasMiddleItems) {
-			items.push({
-				type: 'divider' as const,
-			});
-		}
-
-		items.push({
-
-			// @ts-ignore
-
-			className: 'text-danger',
-			label: Liferay.Language.get('delete'),
-			onClick: async () => {
-				await openCMPModal({
-					center: true,
-					contentComponent: ({
-						closeModal,
-					}: {
-						closeModal: () => void;
-					}) => (
-						<DeleteTaskModal
-							closeModal={closeModal}
-							onSubmit={async () => {
-								const {error} = await deleteTaskById({
-									taskId: String(task.embedded.id),
-								});
-
-								if (!error) {
-									loadData();
-									displayDeleteSuccessToast(
-										task.embedded.title
-									);
-								}
-								else {
-									displayErrorToast(error);
-								}
-								closeModal();
-							}}
-							title={task.embedded.title}
-						/>
-					),
-					size: 'md',
-					status: 'danger',
-				});
-			},
-			symbolLeft: 'trash',
-		});
-	}
-
-	return items;
-}
 
 const TaskCard = React.memo(
 	forwardRef<HTMLDivElement, {isDragging?: boolean; task: ITask}>(
